// src/types/certificationFormData.ts
export interface CertificationFormData {
  id?:                string;
  career?:            string;
  contentAuthor?:     string;
  faculty?:           string;
  evaluatorUsername?: string;
  classroomId?:       number;
  plan?:              string;
  modality?:          string;
  teacher?:           string;
  teacherCode?:       string;
  responsibleDedtef?: string;
  authorityIds?:      number[];
}